nlx-diarize/requirements.txt
kafka-python
pymongo
importlib
scipy
ffmpy
getpass
pydub
numpy
